Added styling to the About Page with Styled Components Library
Created a GlobalStyle component to adjust overall styles. And Created a reusable Style file in the src folder to add styles to specific parts of the About us page. On some components after the arrow function there were some Styled variable created to make minor adjustments to the reusable components